chore: clean the googledrive pkg
