+++ pkg/component/data/googledrive/v0/README.mdx
@@ -5,7 +5,7 @@ draft: false
 description: "Learn about how to set up a VDP Google Drive component https://github.com/instill-ai/instill-core"
 ---
 
-The Google Drive component is a data component that allows users to google Drive is a file storage and synchronization service developed by Google. It allows users to store files in the cloud, synchronize files across devices, and share files..
+The Google Drive component is a data component that allows users to google Drive is a file storage and synchronization service developed by Google. It allows users to store files in the cloud, synchronize files across devices, and share files.
 It can carry out the following tasks:
 - [Read File](#read-file)
 - [Read Folder](#read-folder)
@@ -82,14 +82,14 @@ Read a file content and metadata from Google Drive.
 | Field | Field ID | Type | Note |
 | :--- | :--- | :--- | :--- |
 | Content | `content` | string | Base64 encoded content of the binary file without the `data:[MIME_TYPE];base64,` prefix. Google Sheets will be exported as CSV, Google Docs as PDF, and Google Slides as PDF. If the file is not a Google file, the content will be the same as the original file. |
-| Created time | `created-time` | string | Time when the file was created. Format: `YYYY-MM-DDTHH:MM:SSZ` |
-| ID | `id` | string | ID of the file. |
+| Created time | `created-time` | string | Time when the file was created. Format: `YYYY-MM-DDTHH:MM:SSZ`. |
+| ID | `id` | string | Unique ID of the file. |
 | MD5 checksum | `md5-checksum` | string | MD5 checksum of the file. This reflects every change made to the file on the server, even those not visible to the user. |
-| MIME type | `mime-type` | string | MIME type of the file. |
-| Modified time | `modified-time` | string | Time when the file was last modified. Format: `YYYY-MM-DDTHH:MM:SSZ` |
+| MIME type | `mime-type` | string | MIME type of the file. For example, `application/pdf`, `text/csv`, `image/jpeg`, etc. |
+| Modified time | `modified-time` | string | Time when the file was last modified. Format: `YYYY-MM-DDTHH:MM:SSZ`. |
 | Name | `name` | string | Name of the file. The file extension will be added automatically based on the exported MIME type. For example, Google Sheets will be exported as CSV, Google Docs as PDF, and Google Slides as PDF. If the file is a Google Sheet and the name is `MySheet`, the exported file will be `MySheet.csv`. If the file is not a Google file, the name will be used as is. |
 | Size | `size` | integer | Size of the file in bytes. |
-| Version | `version` | integer | Version of the file. |
+| Version | `version` | integer | Version of the file in Google Drive. |
 | Web Content Link | `web-content-link` | string | Link for downloading the content of the file in a browser. |
 | Web View Link | `web-view-link` | string | Link for opening the file in a relevant Google editor or viewer in a browser. Usually, web view link is same as shared link. |
 </div>
@@ -130,14 +130,14 @@ Read metadata and content of files under the specified folder in Google Drive.
 | Field | Field ID | Type | Note |
 | :--- | :--- | :--- | :--- |
 | Content | `content` | string | Base64 encoded content of the binary file without the `data:[MIME_TYPE];base64,` prefix. Google Sheets will be exported as CSV, Google Docs as PDF, and Google Slides as PDF. If the file is not a Google file, the content will be the same as the original file. |
-| Created time | `created-time` | string | Time when the file was created. Format: `YYYY-MM-DDTHH:MM:SSZ` |
-| ID | `id` | string | ID of the file. |
+| Created time | `created-time` | string | Time when the file was created. Format: `YYYY-MM-DDTHH:MM:SSZ`. |
+| ID | `id` | string | Unique ID of the file. |
 | MD5 checksum | `md5-checksum` | string | MD5 checksum of the file. This reflects every change made to the file on the server, even those not visible to the user. |
-| MIME type | `mime-type` | string | MIME type of the file. |
-| Modified time | `modified-time` | string | Time when the file was last modified. Format: `YYYY-MM-DDTHH:MM:SSZ` |
+| MIME type | `mime-type` | string | MIME type of the file. For example, `application/pdf`, `text/csv`, `image/jpeg`, etc. |
+| Modified time | `modified-time` | string | Time when the file was last modified. Format: `YYYY-MM-DDTHH:MM:SSZ`. |
 | Name | `name` | string | Name of the file. The file extension will be added automatically based on the exported MIME type. For example, Google Sheets will be exported as CSV, Google Docs as PDF, and Google Slides as PDF. If the file is a Google Sheet and the name is `MySheet`, the exported file will be `MySheet.csv`. If the file is not a Google file, the name will be used as is. |
 | Size | `size` | integer | Size of the file in bytes. |
-| Version | `version` | integer | Version of the file. |
+| Version | `version` | integer | Version of the file in Google Drive. |
 | Web Content Link | `web-content-link` | string | Link for downloading the content of the file in a browser. |
 | Web View Link | `web-view-link` | string | Link for opening the file in a relevant Google editor or viewer in a browser. Usually, web view link is same as shared link. |
 </div>
