+++ pkg/component/data/googledrive/v0/main.go
@@ -4,6 +4,7 @@ package googledrive
 import (
 	"context"
 	"fmt"
+	"strings"
 	"sync"
 
 	_ "embed"
@@ -136,4 +137,83 @@ func (c *component) WithOAuthCredentials(s map[string]any) *component {
 	c.instillAIClientID = base.ReadFromGlobalConfig(cfgOAuthClientID, s)
 	c.instillAIClientSecret = base.ReadFromGlobalConfig(cfgOAuthClientSecret, s)
 	return c
+}
+
+// Now, we support the following types of Google Drive links:
+// 1. Folder: https://drive.google
+// 2. File: https://drive.google.com/file/d/
+// 3. Spreadsheet: https://docs.google.com/spreadsheets/d/
+// 4. Document: https://docs.google.com/document/d/
+// 5. Presentation: https://docs.google.com/presentation/d/
+// 6. Colab: https://colab.research.google.com/drive/
+// So, it means the Google Form, Google Map and other types of links are not supported
+func extractUIDFromSharedLink(driveLink string) (string, error) {
+	patterns := map[string]string{
+		"folder":       "/drive/folders/",
+		"file":         "/file/d/",
+		"spreadsheet":  "/spreadsheets/d/",
+		"document":     "/document/d/",
+		"presentation": "/presentation/d/",
+	}
+
+	for _, pattern := range patterns {
+		if strings.Contains(driveLink, pattern) {
+			parts := strings.Split(driveLink, pattern)
+			if len(parts) < 2 {
+				return "", fmt.Errorf("invalid Google Drive link")
+			}
+			// Sample link: https://drive.google.com/drive/folders/xxxxxx?usp=drive_link
+			// Sample link: https://drive.google.com/file/d/xxxxxx/view?usp=drive_link
+			uidParts := strings.SplitN(parts[1], "?", 2)
+			uidParts = strings.SplitN(uidParts[0], "/", 2)
+			return uidParts[0], nil
+		}
+	}
+
+	return "", fmt.Errorf("unrecognized Google Drive link format")
+}
+
+func isFolder(link string) bool {
+	return strings.Contains(link, "/drive/folders/")
+}
+
+func convertDriveFileToComponentFile(driveFile *drive.File) *file {
+	// Google Drive API only can support downloading the binary data.
+	// So, when the file is not binary, we need to export the file as PDF/CSV first.
+	// To make Google Drive Component can seamlessly work with other components, we need to add the file extension to the file name.
+	fileExtension := exportFileExtension(driveFile.MimeType)
+	if fileExtension != "" {
+		driveFile.Name = addFileExtension(driveFile.Name, fileExtension)
+	}
+
+	return &file{
+		ID:             driveFile.Id,
+		Name:           driveFile.Name,
+		CreatedTime:    driveFile.CreatedTime,
+		ModifiedTime:   driveFile.ModifiedTime,
+		Size:           driveFile.Size,
+		MimeType:       driveFile.MimeType,
+		Md5Checksum:    driveFile.Md5Checksum,
+		Version:        driveFile.Version,
+		WebViewLink:    driveFile.WebViewLink,
+		WebContentLink: driveFile.WebContentLink,
+	}
+}
+
+func exportFileExtension(mimeType string) string {
+	switch mimeType {
+	case "application/vnd.google-apps.spreadsheet":
+		return ".csv"
+	case "application/vnd.google-apps.presentation", "application/vnd.google-apps.document":
+		return ".pdf"
+	default:
+		return ""
+	}
+}
+
+func addFileExtension(fileName, Extension string) string {
+	if !strings.HasSuffix(fileName, Extension) {
+		return fileName + Extension
+	}
+	return fileName
 }
