--- pkg/component/data/googledrive/v0/testdata/credentials.json
@@ -1,13 +1,0 @@
-{
-    "installed": {
-      "client_id": "fake_client_id",
-      "project_id": "fake_pjt_id",
-      "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-      "token_uri": "https://oauth2.googleapis.com/token",
-      "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-      "client_secret": "fake_secret",
-      "redirect_uris": [
-        "http://localhost"
-      ]
-    }
-  }
